extract female fashion to its own module
