+++ nodes.py
@@ -10,6 +10,8 @@ from urllib.request import urlopen
 # Import node mappings from separate files
 from .node_defs.person import NODE_CLASS_MAPPINGS as PERSON_CLASS_MAPPINGS
 from .node_defs.person import NODE_DISPLAY_NAME_MAPPINGS as PERSON_DISPLAY_NAME_MAPPINGS
+from .node_defs.female_fashion import NODE_CLASS_MAPPINGS as FASHION_CLASS_MAPPINGS
+from .node_defs.female_fashion import NODE_DISPLAY_NAME_MAPPINGS as FASHION_DISPLAY_NAME_MAPPINGS
 
 # Get the directory where this file is located
 RESOURCES_DIR = os.path.join(os.path.dirname(os.path.realpath(__file__)), "resources")
@@ -108,135 +110,6 @@ def _get_default_character_data():
         "skin_acne_list":["a few blemishes","some acne","pronounced acne"],
         "skin_imperfections_list":["minor imperfections","natural imperfections","pronounced imperfections"]
     }
-
-
-# Character Prompt Builder - Female Fashion Node (Style + Outfit + Accessories)
-class CharacterPromptBuilderFemaleFashion:
-    @classmethod
-    def INPUT_TYPES(s):
-        data = _load_character_data()
-
-        def combo(key, default=None):
-            _list = data.get(key, ["-"]).copy()
-            if '-' not in _list:
-                _list.insert(0, '-')
-            return (_list, {"default": default} if default else {})
-
-        return {
-            "required": {
-                "fashion_aesthetic": combo("fashion_aesthetic_list"),
-                "tops": combo("tops_list"),
-                "tops_color": combo("tops_color_list"),
-                "tops_material": combo("tops_material_list"),
-                "pants": combo("pants_list"),
-                "pants_color": combo("pants_color_list"),
-                "pants_material": combo("pants_material_list"),
-                "dresses": combo("dresses_list"),
-                "dresses_color": combo("dresses_color_list"),
-                "dresses_material": combo("dresses_material_list"),
-                "legs": combo("legs_list"),
-                "legs_color": combo("legs_color_list"),
-                "underwear": combo("underwear_list"),
-                "underwear_color": combo("underwear_color_list"),
-                "underwear_material": combo("underwear_material_list"),
-                "capes": combo("capes_list"),
-                "capes_color": combo("capes_color_list"),
-                "capes_material": combo("capes_material_list"),
-                "hats": combo("hats_list"),
-                "hats_color": combo("hats_color_list"),
-                "womens_suits": combo("womens_suits_list"),
-                "womens_suits_helmet": combo("womens_suits_helmet_list"),
-            },
-            "optional": {
-                "womens_shoes": combo("womens_shoes_list"),
-                "womens_shoe_color": combo("womens_shoe_color_list"),
-                "womens_shoe_material": combo("womens_shoe_material_list"),
-                "womens_gloves": combo("womens_gloves_list"),
-                "womens_gloves_color": combo("womens_gloves_color_list"),
-                "womens_gloves_material": combo("womens_gloves_material_list"),
-                "necklace": combo("necklace_list"),
-                "earrings": combo("earrings_list"),
-                "bracelet": combo("bracelet_list"),
-                "watches": combo("watches_list"),
-                "watches_color": combo("watches_color_list"),
-                "ring": combo("ring_list"),
-                "fingernail_style": combo("fingernail_style_list"),
-                "nail_color": combo("nail_color_list"),
-                "settings_in": ("PM_SETTINGS",),
-                "womens_glasses": combo("womens_glasses_list"),
-                "womens_glasses_color": combo("womens_glasses_color_list"),
-                "stretched_material": (
-                    "BOOLEAN",
-                    {
-                        "default": False,
-                        "label": "Show subtle outline"
-                    }
-                 ),
-                "custom_clothing": (
-                    "STRING",
-                    {
-                        "multiline": True,
-                        "default": "",
-                        "placeholder": "Enter custom outfit"
-                    }
-                ),
-            }
-        }
-
-    RETURN_TYPES = ("PM_SETTINGS",)
-    RETURN_NAMES = ("settings",)
-    FUNCTION = "run"
-    CATEGORY = "CharacterPromptBuilder"
-
-    def run(self,
-            fashion_aesthetic="-",
-            tops="-", tops_color="-", tops_material="-",
-            pants="-", pants_color="-", pants_material="-",
-            dresses="-", dresses_color="-", dresses_material="-",
-            legs="-", legs_color="-",
-            underwear="-", underwear_color="-", underwear_material="-",
-            capes="-", capes_color="-", capes_material="-",
-            hats="-", hats_color="-",
-            womens_suits="-", womens_suits_helmet="-",
-            womens_shoes="-", womens_shoe_color="-", womens_shoe_material="-",
-            womens_gloves="-", womens_gloves_color="-",
-            womens_gloves_material="-",
-            necklace="-", earrings="-", bracelet="-", ring="-",
-            watches="-", watches_color="-",
-            fingernail_style="-", nail_color="-",
-            settings_in=None,
-            womens_glasses="-", womens_glasses_color="-",
-            stretched_material=False,
-            custom_clothing="",
-    ):
-        settings = settings_in.copy() if settings_in else {}
-        settings.update({
-            "fashion_aesthetic": fashion_aesthetic,
-            "tops": tops, "tops_color": tops_color, "tops_material": tops_material,
-            "pants": pants, "pants_color": pants_color, "pants_material": pants_material,
-            "dresses": dresses, "dresses_color": dresses_color, "dresses_material": dresses_material,
-            "legs": legs, "legs_color": legs_color,
-            "underwear": underwear, "underwear_color": underwear_color, "underwear_material": underwear_material,
-            "capes": capes, "capes_color": capes_color, "capes_material": capes_material,
-            "hats": hats, "hats_color": hats_color,
-            "womens_suits": womens_suits,
-            "womens_suits_helmet": womens_suits_helmet,
-            "womens_shoes": womens_shoes, "womens_shoe_color": womens_shoe_color, "womens_shoe_material": womens_shoe_material,
-            "womens_gloves": womens_gloves, "womens_gloves_color": womens_gloves_color,
-            "womens_gloves_material": womens_gloves_material,
-            "necklace": necklace,
-            "earrings": earrings,
-            "bracelet": bracelet,
-            "watches": watches,
-            "watches_color": watches_color,
-            "ring": ring,
-            "fingernail_style": fingernail_style, "nail_color": nail_color,
-            "womens_glasses": womens_glasses,
-            "womens_glasses_color": womens_glasses_color,
-            "stretched_material": stretched_material,
-            "custom_clothing": custom_clothing,
-        })
-        return (settings,)
 
 
 # Character Prompt Builder - Actions Node (Pose/Action)
@@ -1294,14 +1167,14 @@ class CharacterPromptBuilderScene:
 # Node mappings - merge from individual node files
 NODE_CLASS_MAPPINGS = {
     **PERSON_CLASS_MAPPINGS,
-    "Character Prompt Builder Female Fashion": CharacterPromptBuilderFemaleFashion,
+    **FASHION_CLASS_MAPPINGS,
     "Character Prompt Builder Actions": CharacterPromptBuilderActions,
     "Character Prompt Builder Scene": CharacterPromptBuilderScene,
 }
 
 NODE_DISPLAY_NAME_MAPPINGS = {
     **PERSON_DISPLAY_NAME_MAPPINGS,
-    "Character Prompt Builder Female Fashion": "Character Prompt Builder - Female Fashion",
+    **FASHION_DISPLAY_NAME_MAPPINGS,
     "Character Prompt Builder Actions": "Character Prompt Builder - Actions",
     "Character Prompt Builder Scene": "Character Prompt Builder - Scene & Generate",
 }
